Big commit from Week 4 exercises
This commit contains numerous items. It includes the ongoing Word Learning Journal that is being worked on online in Cloudstor. It is up to date from the date of the 23rd of August at 11:30am. This commit also contains the Scoping Exercise II (.pdf and .tex files). This commit also contains Learning Journal Week 4 (.pdf and .tex files) along with all associated .png files.