Proof of Concept - Elaboration I files
This commit contains the .pdf, .tex and .png files that were part of the Elaboration exercise due in Week 5 for FOAR705. This elaboration plan differs from  what was originally proposed in Scoping II because of a change in direction after a meeting I had with an academic mentor. The elaboration plan is specific to the process of collecting articles from the archive of Neues Deutschland, then analysing the corpus with Voyant Tools.